# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample for swagger-maven-plugin",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Kong Chen",
      "url" : "http://kongch.com",
      "email" : "kongchen@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "petstore.swagger.wordnik.com",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/v1/development/get" : {
      "get" : {
        "tags" : [ "v1development" ],
        "summary" : "Retrieves complex data from the server.",
        "description" : "This is only half of what the POST operation does; it only returns data from the server to the client, but does not send any. It's convenient to be called in a web browser.",
        "operationId" : "get",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "apiKey",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A hardcoded consistent ComplexObject.",
            "schema" : {
              "$ref" : "#/definitions/ComplexObject"
            }
          }
        }
      }
    },
    "/v1/development/post" : {
      "post" : {
        "tags" : [ "v1development" ],
        "summary" : "Send various data to the server.",
        "description" : "Ensures that all these data types correctly make it to the server and back. Marshalling of all these types works.",
        "operationId" : "post",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "apiKey",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "The ComplexObject as input.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ComplexObject"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A ComplexObject based on the input.",
            "schema" : {
              "$ref" : "#/definitions/ComplexObject"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "ComplexObject" : {
      "type" : "object",
      "properties" : {
        "string" : {
          "type" : "string",
          "readOnly" : true
        },
        "intNumber" : {
          "type" : "integer",
          "format" : "int32",
          "readOnly" : true
        },
        "doubleNumber" : {
          "type" : "number",
          "format" : "double",
          "readOnly" : true
        },
        "yesOrNo" : {
          "type" : "boolean",
          "readOnly" : true,
          "default" : false
        },
        "color" : {
          "type" : "string",
          "readOnly" : true,
          "enum" : [ "RED", "GREEN", "BLUE" ]
        },
        "geometricalFigure" : {
          "readOnly" : true,
          "$ref" : "#/definitions/GeometricalFigure"
        },
        "optional1" : {
          "type" : "string",
          "readOnly" : true
        },
        "optional2" : {
          "type" : "string",
          "readOnly" : true
        }
      }
    },
    "GeometricalFigure" : {
      "type" : "object",
      "discriminator" : "type",
      "properties" : {
        "color" : {
          "type" : "string",
          "readOnly" : true
        }
      }
    },
...
  }
}